backend/requirements.txt
flask
flask-cors
requests
python-dotenv
Pillow
torch
transformers
opencv-python

